Organize source generation projects to easily manage common code.
