+++ src/SourceGenerators/CommonCode/EmbeddedResource.cs
