+++ src/SourceGenerators/CommonCode/GeneratorExecutionContextExtensions.cs
