+++ src/SourceGenerators/CommonCode/System/Runtime/CompilerServices/IsExternalInit.cs
