+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/AssemblyInfoGenerator.cs
