+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/CSharp.sbntxt
