+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/FSharp.sbntxt
