+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/Model.cs
