+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/ThisAssembly.AssemblyInfo.csproj
@@ -1,0 +1,23 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <PackageId>ThisAssembly.AssemblyInfo</PackageId>
+    <Description>** C# 9.0 ONLY **
+This package generates a static `ThisAssembly.Info` class with public 
+constants exposing the following attribute values generated by default for SDK style projects:
+
+* AssemblyConfigurationAttribute
+* AssemblyCompanyAttribute
+* AssemblyTitleAttribute
+* AssemblyProductAttribute
+
+* AssemblyVersionAttribute
+* AssemblyInformationalVersionAttribute
+* AssemblyFileVersionAttribute
+
+If your project includes these attributes by other means, they will still be emitted properly 
+on the `ThisAssembly.Info` class.
+</Description>
+  </PropertyGroup>
+  
+</Project>
