+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/ThisAssembly.AssemblyInfo.targets
