+++ src/SourceGenerators/ThisAssembly.AssemblyInfo/VisualBasic.sbntxt
