+++ src/SourceGenerators/ThisAssembly.Constants/CSharp.sbntxt
