+++ src/SourceGenerators/ThisAssembly.Constants/ConstantsGenerator.cs
