+++ src/SourceGenerators/ThisAssembly.Constants/Model.cs
