+++ src/SourceGenerators/ThisAssembly.Constants/ThisAssembly.Constants.targets
