+++ src/SourceGenerators/ThisAssembly.Metadata/CSharp.sbntxt
