+++ src/SourceGenerators/ThisAssembly.Metadata/FSharp.sbntxt
