+++ src/SourceGenerators/ThisAssembly.Metadata/MetadataGenerator.cs
