+++ src/SourceGenerators/ThisAssembly.Metadata/Model.cs
