+++ src/SourceGenerators/ThisAssembly.Metadata/ThisAssembly.Metadata.csproj
@@ -1,10 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
     <PackageId>ThisAssembly.Metadata</PackageId>
     <Description>** C# 9.0 ONLY **
 This package generates a static `ThisAssembly.Metadata` class with public 
@@ -35,20 +31,5 @@ C#:
   }
 </Description>
   </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="NuGetizer" Version="0.7.5" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="all" />
-
-    <PackageReference Include="Scriban" Version="5.4.5" Pack="false" IncludeAssets="build" />
-    <PackageReference Include="Microsoft.CSharp" Version="4.5.0" />
-    <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.4" />
-
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.8.0" Pack="false" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ThisAssembly.Prerequisites\ThisAssembly.Prerequisites.csproj" />
-  </ItemGroup>
 
 </Project>
