+++ src/SourceGenerators/ThisAssembly.Metadata/ThisAssembly.Metadata.targets
