+++ src/SourceGenerators/ThisAssembly.Metadata/VisualBasic.sbntxt
