+++ src/SourceGenerators/ThisAssembly.Project/CSharp.sbntxt
