+++ src/SourceGenerators/ThisAssembly.Project/FSharp.sbntxt
