+++ src/SourceGenerators/ThisAssembly.Project/Model.cs
