+++ src/SourceGenerators/ThisAssembly.Project/ProjectPropertyGenerator.cs
