+++ src/SourceGenerators/ThisAssembly.Project/ThisAssembly.Project.props
