+++ src/SourceGenerators/ThisAssembly.Project/ThisAssembly.Project.targets
