+++ src/SourceGenerators/ThisAssembly.Project/VisualBasic.sbntxt
