+++ src/SourceGenerators/ThisAssembly.Strings/CSharp.sbntxt
