+++ src/SourceGenerators/ThisAssembly.Strings/FSharp.sbntxt
