+++ src/SourceGenerators/ThisAssembly.Strings/Model.cs
