+++ src/SourceGenerators/ThisAssembly.Strings/StringsGenerator.cs
