+++ src/SourceGenerators/ThisAssembly.Strings/ThisAssembly.Strings.cs
