+++ src/SourceGenerators/ThisAssembly.Strings/ThisAssembly.Strings.props
