+++ src/SourceGenerators/ThisAssembly.Strings/ThisAssembly.Strings.targets
