+++ src/SourceGenerators/ThisAssembly.Strings/VisualBasic.sbntxt
