--- src/ThisAssembly.Project/ThisAssembly.Project.csproj
@@ -1,57 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>ThisAssembly.Project</PackageId>
-    <Description>** C# 9.0 ONLY **
-This package generates a static `ThisAssembly.Project` class with public 
-constants exposing project properties that have been opted into this mechanism by adding 
-them as `ProjectProperty` MSBuild items in project file, such as:
-
-    &lt;PropertyGroup&gt;
-      &lt;Foo&gt;Bar&lt;/Foo&gt;
-    &lt;/PropertyGroup&gt;
-
-    &lt;ItemGroup&gt;
-      &lt;ProjectProperty Include="Foo" /&gt;
-    &lt;/ItemGroup&gt;
-
-A corresponding `ThisAssembly.Project.Foo` constant with the value `Bar` is provided. 
-
-Generated code:
-C#:
-
-  partial class ThisAssembly
-  {
-      public static partial class Project
-      {
-          public const string Foo = "Bar";
-      }
-  }
-
-</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\ThisAssembly.Metadata\HashCode.cs" Link="HashCode.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="NuGetizer" Version="0.7.5" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="all" />
-
-    <PackageReference Include="Scriban" Version="5.4.5" Pack="false" IncludeAssets="build" />
-    <PackageReference Include="Microsoft.CSharp" Version="4.5.0" />
-    <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.4" />
-
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.8.0" Pack="false" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ThisAssembly.Prerequisites\ThisAssembly.Prerequisites.csproj" />
-  </ItemGroup>
-
-</Project>
